Move nth_prime, is_prime from lib.rs to number_theory.rs
The functions is_prime and nth_prime are not top level library
functions, they better belong in number_theory.rs.
