+++ tests/interpreter_tests/math/number_theory.rs
@@ -3901,6 +3901,7 @@ mod arithmetic_geometric_mean {
 
 mod random_prime {
   use super::*;
+  use woxi::functions::math_ast::is_prime;
 
   #[test]
   fn with_max() {
@@ -3908,7 +3909,7 @@ mod random_prime {
     interpret("SeedRandom[42]").unwrap();
     let result: i128 = interpret("RandomPrime[100]").unwrap().parse().unwrap();
     assert!(result >= 2 && result <= 100);
-    assert!(woxi::is_prime(result as usize));
+    assert!(is_prime(result as usize));
   }
 
   #[test]
@@ -3917,7 +3918,7 @@ mod random_prime {
     let result: i128 =
       interpret("RandomPrime[{10, 30}]").unwrap().parse().unwrap();
     assert!(result >= 10 && result <= 30);
-    assert!(woxi::is_prime(result as usize));
+    assert!(is_prime(result as usize));
   }
 
   #[test]
